StockReportv2/requirements.txt
pandas>=2.1
numpy>=1.26
yfinance>=0.2.40
scipy>=1.11
xlsxwriter>=3.2
pyyaml>=6.0
